telegram-bot/requirements.txt
python-telegram-bot==21.0.1
aiohttp>=3.9.0